delete sharedStateController
